refactor: clean up and standardise file names
