+++ src/utils/index.ts
@@ -1,5 +1,5 @@
-export * from "./computeTitleEscrowAddress";
-export * from "./encodeInitParams";
-export * from "./isSupportedTitleEscrowFactory";
-export * from "./getEventFromTransaction";
-export * from "./computeInterfaceId";
+export * from "./compute-title-escrow-address";
+export * from "./encode-init-params";
+export * from "./is-supported-title-escrow-factory";
+export * from "./get-event-from-transaction";
+export * from "./compute-interface-id";
